+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## v0.17.1
 
-- Fix default canvas width to fix an issue on high-dpi monitors that do not style the canvas element (#41)
+- Add default canvas width/height (`'auto'`) to fix an issue on high-dpi monitors that do not style the canvas element (#41)
+- Add resize handler so when width/height is set to `'auto'` the canvas element is automatically resized on `resize` and `orientationchange` events.
 
 ## v0.17.0